add attendees route, reinstated date field to event model and routes -

+++ server/routes/events.js
@@ -11,6 +11,7 @@ const db = require('../models');
 router.post('/', async (req, res) => {
 	let {
 		location,
+		date,
 		startTime,
 		endTime,
 		languages,
@@ -34,6 +35,7 @@ router.post('/', async (req, res) => {
 		const event = new db.Event({
 			createdBy: req.user._id,
 			location,
+			date: new Date(date),
 			startTime: new Date(startTime),
 			endTime: new Date(endTime),
 			languages,
@@ -54,7 +56,12 @@ router.post('/', async (req, res) => {
 
 router.get('/', async (req, res) => {
 	try {
-		const event = await db.Event.find();
+		const event = await db.Event.find()
+			.populate('attendees.attendee')
+			.populate('matches.user1')
+			.populate('matches.user2')
+			.populate('matches.user3')
+			.populate('matches.user4');
 		res.json(event);
 	} catch (err) {
 		console.error(err.message);
@@ -69,7 +76,12 @@ router.get('/:id', async (req, res) => {
 	try {
 		const event = await db.Event.find({
 			_id: req.params.id,
-		});
+		})
+			.populate('attendees.attendee')
+			.populate('matches.user1')
+			.populate('matches.user2')
+			.populate('matches.user3')
+			.populate('matches.user4');
 		res.json(event);
 	} catch (err) {
 		console.error(err.message);
@@ -81,10 +93,11 @@ router.get('/:id', async (req, res) => {
 // @desc    Allows Admin to update event
 
 router.put('/:id', async (req, res) => {
-	const { location, startTime, endTime, languages, levels } = req.body;
+	const { location, date, startTime, endTime, languages, levels } = req.body;
 	try {
 		const event = {};
 		if (location) event.location = location;
+		if (date) event.date = date;
 		if (startTime) event.startTime = startTime;
 		if (endTime) event.endTime = endTime;
 		if (languages) event.languages = languages;
@@ -94,7 +107,7 @@ router.put('/:id', async (req, res) => {
 
 		if (user.admin !== true) {
 			return res.status(401).json({
-				msg: 'You are not authorized to edit foods in this event.',
+				msg: 'You are not authorized to edit this event.',
 			});
 		}
 
@@ -108,22 +121,26 @@ router.put('/:id', async (req, res) => {
 		res.status(500).send('Server Error');
 	}
 });
-// @route   UPDATE /attendees/:id
-// @desc    Allows Admin to update event
-router.put('/attendees/:id', async (req, res) => {
-	const { attendees } = req.body;
+// @route   UPDATE attendees/:userId/:eventId
+// @desc    Adds attendees to event once they sign in.
+router.put('/attendees/:userId/:eventId', async (req, res) => {
 	try {
+		// check to make sure user making updates has admin rights.
 		let user = await db.User.findOne({ _id: req.user._id });
-
 		if (user.admin !== true) {
 			return res.status(401).json({
-				msg: 'You are not authorized to edit foods in this event.',
+				msg: 'You are not authorized to edit this event.',
 			});
 		}
 
+		const attendee = await db.User.findOne({ _id: req.params.userId });
 		await db.Event.findOneAndUpdate(
-			{ _id: req.params.id },
-			{ $push: attendees }
+			{ _id: req.params.eventId },
+			{
+				$push: {
+					attendees: { attendee },
+				},
+			}
 		);
 		res.send('Your event was updated!');
 	} catch (err) {
@@ -131,14 +148,47 @@ router.put('/attendees/:id', async (req, res) => {
 		res.status(500).send('Server Error');
 	}
 });
-// @route   UPDATE /matches/:id
+// @route   UPDATE /matches/:eventId
 // @desc    Allows Admin to update event
+router.put('/matches/:eventId', async (req, res) => {
+	const { matches } = req.body;
+	// console.log(matches);
+	// console.log(req.params.eventId);
+	try {
+		// check to make sure user making updates has admin rights.
+		let user = await db.User.findOne({ _id: req.user._id });
+		if (user.admin !== true) {
+			return res.status(401).json({
+				msg: 'You are not authorized to edit this event.',
+			});
+		}
+		await db.Event.findOneAndUpdate(
+			{ _id: req.params.eventId },
+			{
+				$push: {
+					matches: matches,
+				},
+			}
+		);
+		res.send('Your event was updated!');
+	} catch (err) {
+		console.error(err.message);
+		res.status(500).send('Server Error');
+	}
+});
 
 // @route   DELETE delete/:id - [works 2/12]
 // @desc    Route to delete whole event if createdby user = req.user._id
 
 router.delete('/:id', async (req, res) => {
 	try {
+		// check to make sure user making updates has admin rights.
+		let user = await db.User.findOne({ _id: req.user._id });
+		if (user.admin !== true) {
+			return res.status(401).json({
+				msg: 'You are not authorized to edit this event.',
+			});
+		}
 		await db.Event.findOneAndDelete({
 			_id: req.params.id,
 		});
